Add CheckStyle and fix CheckStyle violations
