--- src/main/java/Bob/enums/CommandType.java
@@ -1,5 +1,0 @@
-package Bob.enums;
-
-public enum CommandType {
-    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, FIND, INVALID
-}
