--- src/main/java/Bob/exception/BobCorruptFileException.java
@@ -1,7 +1,0 @@
-package Bob.exception;
-
-public class BobCorruptFileException extends BobException{
-    public BobCorruptFileException(String errorMessage) {
-        super("Your file may be corrupted :(");
-    }
-}
