--- src/main/java/Bob/exception/BobException.java
@@ -1,7 +1,0 @@
-package Bob.exception;
-
-public class BobException extends Exception {
-    public BobException(String errorMessage) {
-        super(errorMessage);
-    }
-}
