--- src/main/java/Bob/exception/BobInvalidCommandException.java
@@ -1,11 +1,0 @@
-package Bob.exception;
-
-public class BobInvalidCommandException extends BobException {
-    public BobInvalidCommandException() {
-        super("Invalid command format chosen");
-    }
-
-    public BobInvalidCommandException(String message) {
-        super(message);
-    }
-}
