--- src/main/java/Bob/exception/BobInvalidTaskNumberException.java
@@ -1,11 +1,0 @@
-package Bob.exception;
-
-public class BobInvalidTaskNumberException extends BobException {
-    public BobInvalidTaskNumberException() {
-        super("Index chosen for task is invalid");
-    }
-
-    public BobInvalidTaskNumberException(String message) {
-        super(message);
-    }
-}
