--- src/main/java/Bob/exception/BobMissingArgumentException.java
@@ -1,11 +1,0 @@
-package Bob.exception;
-
-public class BobMissingArgumentException extends BobException{
-    public BobMissingArgumentException() {
-        super("You are missing an argument!");
-    }
-
-    public BobMissingArgumentException(String message) {
-        super(message);
-    }
-}
