+++ src/main/java/bob/enums/CommandType.java
@@ -1,0 +1,8 @@
+package bob.enums;
+
+/**
+ * Enumerations for types of commands available to be executed.
+ */
+public enum CommandType {
+    BYE, LIST, MARK, UNMARK, TODO, DEADLINE, EVENT, DELETE, FIND, INVALID
+}
